Make sure macros are defined during native compilation.
Emacs 28 supports native compilation of byte-compiled modules.
Native compilation is performed asynchronously in the background, and
modules may be compiled in a different order than the order in which
they are imported when running mew.  This commit adds a number
of (require) forms to make sure all macros are defined when they are
used.