Separately check if stunnel supports 'foreground', 'pid' and 'syslog' options
While investigating the problem that Mew doesn't work well with
version 5.68 of stunnel for Windows, I noticed that it supports
'foreground' and 'syslog' options but doesn't supports 'pid'
option. So change mew-ssl.el so it checks if stunnel supports each of
3 options separately and only adds setting of supported one to
configuration file.

While I'm here, add documentation string to mew-ssl-libwrap and
mew-ssl-checkhost.