Fix stopping during greeting on some POPs servers
Patch by Hiroki Sato
https://groups.google.com/g/mew-ja/c/Z7r8K5GbKuc/m/ps_N-_gmAAAJ

When using GnuTLS, POP greetings are not received
because GnuTLS requires a client-initiated command
after the session is established or upgraded to use TLS.
Therefore, a dummy CAPA was sent,
and the response was used as a substitute for the greeting.

However, some POPs servers (e.g., secure.plala.or.jp:995)
split CAPA responses into multiple packets.
Mew stopped greeting unless the substituted CAPA response
was a single packet.

This commit fixes it by removing the sending of the dummy CAPA
and passing dummy greetings to the process filter.